Refactor new bibDetails methods to util filed and add comments
